+++ radiospectra/spectrogram/sources/__init__.py
@@ -8,6 +8,7 @@ Each mission should have its own file with one or more classes defined.
 from ..spectrogram_factory import Spectrogram  # NOQA
 from .callisto import *  # NOQA
 from .eovsa import *  # NOQA
+from .ilofar357 import *  # NOQA
 from .psp_rfs import *  # NOQA
 from .rpw import *  # NOQA
 from .rstn import *  # NOQA
@@ -21,4 +22,5 @@ __all__ = [
     "EOVSASpectrogram",
     "RSTNSpectrogram",
     "RPWSpectrogram",
+    "ILOFARMode357Spectrogram",
 ]