+++ radiospectra/spectrogram/spectrogram_factory.py
@@ -16,6 +16,7 @@ import astropy.units as u
 from astropy.io import fits
 from astropy.io.fits import Header
 from astropy.time import Time
+from sunpy import log
 from sunpy.data import cache
 from sunpy.net import attrs as a
 from sunpy.time import parse_time
@@ -32,6 +33,7 @@ from sunpy.util.util import expand_list
 
 from radiospectra.exceptions import NoSpectrogramInFileError, SpectraMetaValidationError
 from radiospectra.spectrogram.spectrogrambase import GenericSpectrogram
+from radiospectra.utils import subband_to_freq
 
 SUPPORTED_ARRAY_TYPES = (np.ndarray,)
 try:
@@ -236,7 +238,7 @@ class SpectrogramFactory(BasicRegistrationFactory):
         extensions = file.suffixes
         first_extension = extensions[0].lower()
         if first_extension == ".dat":
-            return [self._read_dat(file)]
+            return self._read_dat(file)
         elif first_extension in (".r1", ".r2"):
             return [self._read_idl_sav(file, instrument="waves")]
         elif first_extension == ".cdf":
@@ -276,6 +278,49 @@ class SpectrogramFactory(BasicRegistrationFactory):
             meta["times"] = meta["start_time"] + times
             meta["end_time"] = meta["start_time"] + times[-1]
             return data, meta
+        elif "bst" in file.name:
+            subbands = (np.arange(54, 454, 2), np.arange(54, 454, 2), np.arange(54, 230, 2))
+            num_subbands = 488
+
+            data = np.fromfile(file)
+            polarisation = file.stem[-1]
+
+            num_times = data.shape[0] / num_subbands
+            if not num_times.is_integer():
+                log.warning("BST file seems incomplete dropping incomplete frequencies")
+                num_times = np.floor(num_times).astype(int)
+                truncate = num_times * num_subbands
+                data = data[:truncate]
+            data = data.reshape(-1, num_subbands).T  # (Freq x Time).T = (Time x Freq)
+            dt = np.arange(num_times) * 1 * u.s
+            start_time = Time.strptime(file.name.split("_bst")[0], "%Y%m%d_%H%M%S")
+            times = start_time + dt
+
+            obs_mode = (3, 5, 7)
+
+            freqs = [subband_to_freq(sb, mode) for sb, mode in zip(subbands, obs_mode)]
+
+            # 1st 200 sbs mode 3, next 200 sbs mode 5, last 88 sbs mode 7
+            spec = {0: data[:200, :], 1: data[200:400, :], 2: data[400:, :]}
+            data_header_pairs = []
+            for i in range(3):
+                meta = {
+                    "instrument": "ILOFAR",
+                    "observatory": "Birr (IE613)",
+                    "start_time": times[0],
+                    "mode": obs_mode[i],
+                    "wavelength": a.Wavelength(freqs[i][0], freqs[i][-1]),
+                    "freqs": freqs[i],
+                    "times": times,
+                    "end_time": times[-1],
+                    "detector": "ILOFAR",
+                    "polarisation": polarisation,
+                }
+
+                data_header_pairs.append((spec[i], meta))
+            return data_header_pairs
+        else:
+            raise ValueError(f"File {file} not supported.")
 
     @staticmethod
     def _read_srs(file):